caas-backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Stripe Integration
stripe==7.11.0

# AI Providers
openai==1.10.0
google-generativeai==0.3.2
tiktoken==0.5.2

# Redis (for caching & coordination)
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.6

# Development
ipython==8.20.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
